Atualização das dependências de testcontainer e restassured e limitação dos testes de testcontainer para não rodar em ambiente windows devido a limitação da dependência
