+++ src/test/java/one/digitalinnovation/parking/controller/MockMvc/ParkedControllerTest.java
@@ -1,0 +1,5 @@
+package one.digitalinnovation.parking.controller.MockMvc;
+
+public class ParkedControllerTest {
+
+}
