+++ src/test/java/one/digitalinnovation/parking/controller/MockMvc/VehicleControllerTest.java
@@ -1,0 +1,5 @@
+package one.digitalinnovation.parking.controller.MockMvc;
+
+public class VehicleControllerTest {
+
+}
