--- src/test/java/one/digitalinnovation/parking/controller/VehicleControllerTest.java
@@ -1,54 +1,0 @@
-package one.digitalinnovation.parking.controller;
-
-import org.hamcrest.Matchers;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-
-import io.restassured.RestAssured;
-import one.digitalinnovation.parking.controller.dto.VehicleCreateDTO;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class VehicleControllerTest extends AbstractContainerBase {
-
-    @LocalServerPort
-    private int randomPort;
-
-    @BeforeEach
-    public void setUpTest() {
-        RestAssured.port = randomPort;
-    }
-
-    @Test
-    void whenFindAllThenCheckResult() {
-        RestAssured.given()
-                .auth().basic("user", "12345")
-                .when()
-                .get("/parking")
-                .then()
-                .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void whenCreateThenCheckIsCreated() {
-        var createDTO = new VehicleCreateDTO();
-        createDTO.setColor("AMARELO");
-        createDTO.setLicense("WRT-5555");
-        createDTO.setModel("BRASILIA");
-        createDTO.setState("SP");
-
-        RestAssured.given()
-                .when()
-                .auth().basic("user", "12345")
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(createDTO)
-                .post("/parking")
-                .then()
-                .statusCode(HttpStatus.CREATED.value())
-                .body("license", Matchers.equalTo("WRT-5555"))
-                .body("color", Matchers.equalTo("AMARELO"));
-    }
-}
